Reverted non-offscreen impl back to variant without use of transferhandlers
